backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
google-generativeai
pydub
librosa
scikit-learn
numpy
PyPDF2
python-docx
requests
beautifulsoup4
werkzeug
bcrypt
soundfile
webrtcvad
pymongo
python-dotenv
